fix: resolve relationship compatibility, changes tab visibility, custom model support, morph maps, and policy typehints
- fix(#34): resolve relation dynamically in ActivitiesRelationManager
- fix(#33, #30): fix changes tab visibility and TypeError when changes contain array/object values
- fix(#32): cast intensity level calculation to integer for strict matching in heatmap widget
- feat(#31): respect custom activity model from Spatie config in resource, widgets, and policy registration
- fix(#29): support morph maps in subject resource URLs, restore actions, and filters
- fix(#27): typehint Spatie Activity contract in ActivityPolicy to prevent TypeError with custom models

+++ resources/views/widgets/activity-heatmap.blade.php
@@ -60,7 +60,7 @@
                             $currentDate = $startDate->copy()->addWeeks($week)->addDays($day);
                             $dateString = $currentDate->toDateString();
                             $count = $data[$dateString] ?? 0;
-                            $intensity = $count > 0 ? ceil(($count / $max) * 4) : 0;
+                            $intensity = $count > 0 ? (int) ceil(($count / $max) * 4) : 0;
                             
                             // Universal colors (subtler empty state)
                             $bg = match ($intensity) {
